Move resolve_site to utils
